requirements.txt
lorem==0.1.1
PyYAML==6.0.2
selenium==4.27.1
